Refactor opt-remarks: make gcc dump remarks to file, separate file handling to ClangCompiler and GccCompiler, make DefaultCompiler inherit GccCompiler, move some tests around
