+++ lib/compilers/clang.ts
@@ -24,6 +24,8 @@
 
 import fs from 'node:fs';
 import path from 'node:path';
+import {parseAllDocuments} from 'yaml';
+import {logger} from '../logger.js';
 
 import _ from 'underscore';
 
@@ -310,6 +312,51 @@ export class ClangCompiler extends BaseCompiler {
         return this.llvmIr.isLlvmIr(deviceAsm)
             ? this.llvmIr.process(deviceAsm, filters)
             : this.asm.process(deviceAsm, filters);
+    }
+
+    override processRawOptRemarks(buffer: string, compileFileName = ''): OptRemark[] {
+        const output: OptRemark[] = [];
+        const remarksSet: Set<string> = new Set<string>();
+        const remarks: any = parseAllDocuments(buffer);
+
+        const DisplayOptInfo = (optInfo: OptRemark) => {
+            let displayString = optInfo.Args.reduce((acc, x) => {
+                let inc = '';
+                for (const [key, value] of Object.entries(x)) {
+                    if (key === 'DebugLoc') {
+                        if (value['Line'] !== 0) {
+                            inc += ' (' + value['Line'] + ':' + value['Column'] + ')';
+                        }
+                    } else {
+                        inc += value;
+                    }
+                }
+                return acc + inc;
+            }, '');
+
+            displayString = displayString.replaceAll('\n', ' ').replaceAll('\r', ' ');
+            return displayString;
+        };
+
+        for (const doc of remarks) {
+            if (doc.errors !== undefined && doc.errors.length > 0) {
+                logger.warn('YAMLParseError: ' + JSON.stringify(doc.errors[0]));
+                continue;
+            }
+
+            const opt = doc.toJS();
+            if (!opt.DebugLoc || !opt.DebugLoc.File || !opt.DebugLoc.File.includes(compileFileName)) continue;
+
+            const strOpt = JSON.stringify(opt);
+            if (!remarksSet.has(strOpt)) {
+                remarksSet.add(strOpt);
+                opt.optType = doc.contents.tag.substring(1); // remove leading '!'
+                opt.displayString = DisplayOptInfo(opt);
+                output.push(opt as OptRemark);
+            }
+        }
+
+        return output;
     }
 }
