+++ lib/compilers/rust.ts
@@ -249,7 +249,7 @@ export class RustCompiler extends BaseCompiler {
         return opts;
     }
 
-    override getOptYamlPath(dirPath: string, outputFilebase: string): string {
+    override getOptFilePath(dirPath: string, outputFilebase: string): string {
         // Find a file in dirPath that ends with codegen.opt.yaml, and return it
         // A bit of a hack, but it works for now
         const files = readdirSync(dirPath);
